project_root/requirements.txt
# In requirements.txt
pandas
numpy
requests
TA-Lib
websocket-client
protobuf
scikit-learn 
numba 